pyproject.toml
[project]
name = "mlops-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.3",
    "mlflow[extras]>=3.1.1",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]
ml = [
    "optuna>=4.4.0",
    "sentence-transformers>=5.0.0",
]
